requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.36.1
botocore==1.36.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
google-auth==2.37.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
openai==1.59.8
packaging==24.2
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
s3transfer==0.11.1
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.8.2